# openapi/openapi.yaml
openapi: 3.0.3
servers:
  - url: http://localhost:3000
info:
  title: Marketplace API
  version: 1.0.0
security: [ ]
paths:
  /products:
    get:
      operationId: getAllProducts
      summary: Get your all products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          description: Opaque pagination token from the previous response's next_cursor.
          schema:
            type: string
      responses:
        '200':
          description: Your all products are here
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  next_cursor:
                    type: string
                    nullable: true
        '404':
          description: There is something wrong with your products
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /products/{id}:
    get:
      operationId: getProduct
      summary: Get your product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The product is found
        '404':
          description: The current product is unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /orders:
    get:
      operationId: getAllOrders
      summary: Get your all orders
      responses:
        '200':
          description: Your all orders are here
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  next_cursor:
                    type: string
                    nullable: true
        '404':
          description: There is something wrong with your orders
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: createOrder
      summary: Create a new order
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Key that is generated by client to avoid double same requests
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
      responses:
        '201':
          description: The order is created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  items:
                    type: array
        '400':
          description: Invalid order data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /orders/{id}:
    get:
      operationId: getOrder
      summary: Get your order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The order is found
        '404':
          description: The current order is unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    Problem:
      type: object
      required: [ type, title, status, detail, instance ]
      properties:
        type:
          type: string
          description: URI типу помилки
        title:
          type: string
          description: Коротка назва помилки
        status:
          type: integer
          description: HTTP-код
        detail:
          type: string
          description: Що саме пішло не так
        instance:
          type: string
          description: Посилання на конкретний випадок